feat: add the horizontal moving to the basic operations provided by the inverter device

+++ VerticalWarehouses/InverterDriver/Source/ParameterID.cs
@@ -7,19 +7,33 @@
     public enum ParameterID
     {
         CONTROL_WORD_PARAM = 410,
+
         HOMING_CREEP_SPEED_PARAM = 1133,
+
         HOMING_FAST_SPEED_PARAM = 1132,
+
         HOMING_MODE_PARAM = 1130,
+
         HOMING_OFFSET_PARAM = 1131,
+
         POSITION_ACCELERATION_PARAM = 1457,
+
         POSITION_DECELERATION_PARAM = 1458,
+
         POSITION_TARGET_POSITION_PARAM = 1455,
+
         POSITION_TARGET_SPEED_PARAM = 1456,
+
         SET_OPERATING_MODE_PARAM = 1454,
+
         STATUS_WORD_PARAM = 411,
+
         ACTUAL_POSITION_SHAFT = 1108,
+
         STATUS_DIGITAL_SIGNALS = 250,
+
         CONTROL_MODE_PARAM = 412,   // Local/Remote: 1-State Machine
+
         ANALOG_IC_PARAM = 457
     }
 
@@ -60,7 +74,7 @@
                 case ParameterID.POSITION_ACCELERATION_PARAM: type = ValueDataType.Float; break;
                 case ParameterID.POSITION_DECELERATION_PARAM: type = ValueDataType.Float; break;
                 case ParameterID.POSITION_TARGET_POSITION_PARAM: type = ValueDataType.Int32; break;
-                case ParameterID.POSITION_TARGET_SPEED_PARAM: type = ValueDataType.Float; break;
+                case ParameterID.POSITION_TARGET_SPEED_PARAM: type = ValueDataType.Int32; break;
                 case ParameterID.SET_OPERATING_MODE_PARAM: type = ValueDataType.Int16; break;
                 case ParameterID.STATUS_WORD_PARAM: type = ValueDataType.UInt16; break;
                 case ParameterID.STATUS_DIGITAL_SIGNALS: type = ValueDataType.Int16; break;
